+++ src/webapp-builtins/export/EventFan.js
@@ -95,7 +95,7 @@ export class EventFan extends BaseService {
 
   /** @override */
   static _impl_configClass() {
-    return class Config extends super.prototype.constructor.CONFIG_CLASS {
+    return class Config extends super.prototype.constructor.configClass {
       // @defaultConstructor
 
       /**