+++ src/webapp-builtins/export/MemoryMonitor.js
@@ -136,7 +136,7 @@ export class MemoryMonitor extends TemplThreadComponent('MemoryThread', BaseServ
 
   /** @override */
   static _impl_configClass() {
-    return class Config extends super.prototype.constructor.CONFIG_CLASS {
+    return class Config extends super.prototype.constructor.configClass {
       // @defaultConstructor
 
       /**