+++ src/webapp-builtins/export/PathRouter.js
@@ -92,7 +92,7 @@ export class PathRouter extends BaseApplication {
 
   /** @override */
   static _impl_configClass() {
-    return class Config extends super.prototype.constructor.CONFIG_CLASS {
+    return class Config extends super.prototype.constructor.configClass {
       // @defaultConstructor
 
       /**