+++ src/webapp-builtins/export/ProcessIdFile.js
@@ -190,7 +190,7 @@ export class ProcessIdFile extends TemplThreadComponent('FileThread', BaseFileSe
 
   /** @override */
   static _impl_configClass() {
-    return class Config extends super.prototype.constructor.CONFIG_CLASS {
+    return class Config extends super.prototype.constructor.configClass {
       // @defaultConstructor
 
       /**