+++ src/webapp-builtins/export/Redirector.js
@@ -68,7 +68,7 @@ export class Redirector extends BaseApplication {
 
   /** @override */
   static _impl_configClass() {
-    return class Config extends super.prototype.constructor.CONFIG_CLASS {
+    return class Config extends super.prototype.constructor.configClass {
       // @defaultConstructor
 
       /**