+++ src/webapp-builtins/export/RequestDelay.js
@@ -51,7 +51,7 @@ export class RequestDelay extends BaseApplication {
 
   /** @override */
   static _impl_configClass() {
-    return class Config extends super.prototype.constructor.CONFIG_CLASS {
+    return class Config extends super.prototype.constructor.configClass {
       // @defaultConstructor
 
       /**