+++ src/webapp-builtins/export/SimpleResponse.js
@@ -108,7 +108,7 @@ export class SimpleResponse extends BaseApplication {
 
   /** @override */
   static _impl_configClass() {
-    return class Config extends super.prototype.constructor.CONFIG_CLASS {
+    return class Config extends super.prototype.constructor.configClass {
       // @defaultConstructor
 
       /**