+++ src/webapp-core/export/NetworkHost.js
@@ -115,7 +115,7 @@ export class NetworkHost extends BaseComponent {
 
   /** @override */
   static _impl_configClass() {
-    return class Config extends super.prototype.constructor.CONFIG_CLASS {
+    return class Config extends super.prototype.constructor.configClass {
       // @defaultConstructor
 
       /**
@@ -211,7 +211,7 @@ export class NetworkHost extends BaseComponent {
    * Makes the parameters for a newly-generated self-signed certificate and
    * corresponding key.
    *
-   * @param {NetworkHost.CONFIG_CLASS} config Parsed configuration item.
+   * @param {NetworkHost.configClass} config Parsed configuration item.
    * @returns {{certificate: string, privateKey: string}} The parameters.
    */
   static async #makeSelfSignedParameters(config) {