+++ src/webapp-util/private/BaseFilePreserver.js
@@ -23,7 +23,7 @@ export class BaseFilePreserver {
   /**
    * Configuration to use.
    *
-   * @type {BaseFileService.CONFIG_CLASS}
+   * @type {BaseFileService.configClass}
    */
   #config;
 
@@ -65,11 +65,11 @@ export class BaseFilePreserver {
   /**
    * Constructs an instance.
    *
-   * @param {BaseFileService.CONFIG_CLASS} config Configuration to use.
+   * @param {BaseFileService.configClass} config Configuration to use.
    * @param {?IntfLogger} logger Logger to use, or `null` to not do any logging.
    */
   constructor(config, logger) {
-    this.#config = MustBe.instanceOf(config, BaseFileService.CONFIG_CLASS);
+    this.#config = MustBe.instanceOf(config, BaseFileService.configClass);
     this.#logger = logger?.saver ?? null;
   }